Refactor `bevy_render_3d::atmosphere`
