+++ crates/bevy_render_3d/src/atmosphere/plugin/shaders/aerial_view_lut.wgsl
