+++ crates/bevy_render_3d/src/atmosphere/plugin/shaders/bruneton_functions.wgsl
