+++ crates/bevy_render_3d/src/atmosphere/plugin/shaders/functions.wgsl
