+++ crates/bevy_render_3d/src/atmosphere/plugin/shaders/multiscattering_lut.wgsl
