+++ crates/bevy_render_3d/src/atmosphere/plugin/shaders/sky_view_lut.wgsl
