+++ crates/bevy_render_3d/src/atmosphere/plugin/systems.rs
@@ -1,0 +1,324 @@
+use bevy_core_pipeline::core_3d::Camera3d;
+use bevy_ecs::{
+    entity::Entity,
+    query::{Changed, With},
+    system::{Commands, Query, Res, ResMut},
+};
+use bevy_math::{Mat4, Vec3};
+use bevy_render::{
+    camera::Camera,
+    extract_component::ComponentUniforms,
+    render_resource::{
+        BindGroupEntries, Extent3d, PipelineCache, SpecializedRenderPipelines, TextureDescriptor,
+        TextureDimension, TextureFormat, TextureUsages,
+    },
+    renderer::{RenderDevice, RenderQueue},
+    texture::TextureCache,
+    view::{ExtractedView, Msaa, ViewDepthTexture, ViewUniforms},
+};
+
+use crate::{
+    resources::{
+        AtmosphereBindGroups, AtmosphereSamplers, AtmosphereTextures, AtmosphereTransform,
+        AtmosphereTransforms, AtmosphereTransformsOffset, RenderSkyPipelineId,
+        RenderSkyPipelineKey,
+    },
+    Atmosphere, AtmosphereSettings, LightMeta,
+};
+
+use super::{AtmosphereBindGroupLayouts, RenderSkyBindGroupLayouts};
+
+pub fn configure_camera_depth_usages(
+    mut cameras: Query<&mut Camera3d, (Changed<Camera3d>, With<Atmosphere>)>,
+) {
+    for mut camera in &mut cameras {
+        camera.depth_texture_usages.0 |= TextureUsages::TEXTURE_BINDING.bits();
+    }
+}
+
+pub fn queue_render_sky_pipelines(
+    views: Query<(Entity, &Camera, &Msaa), With<Atmosphere>>,
+    pipeline_cache: Res<PipelineCache>,
+    layouts: Res<RenderSkyBindGroupLayouts>,
+    mut specializer: ResMut<SpecializedRenderPipelines<RenderSkyBindGroupLayouts>>,
+    mut commands: Commands,
+) {
+    for (entity, camera, msaa) in &views {
+        let id = specializer.specialize(
+            &pipeline_cache,
+            &layouts,
+            RenderSkyPipelineKey {
+                msaa_samples: msaa.samples(),
+                hdr: camera.hdr,
+            },
+        );
+        commands.entity(entity).insert(RenderSkyPipelineId(id));
+    }
+}
+
+pub fn prepare_atmosphere_textures(
+    views: Query<(Entity, &AtmosphereSettings), With<Atmosphere>>,
+    render_device: Res<RenderDevice>,
+    mut texture_cache: ResMut<TextureCache>,
+    mut commands: Commands,
+) {
+    for (entity, lut_settings) in &views {
+        let transmittance_lut = texture_cache.get(
+            &render_device,
+            TextureDescriptor {
+                label: Some("transmittance_lut"),
+                size: Extent3d {
+                    width: lut_settings.transmittance_lut_size.x,
+                    height: lut_settings.transmittance_lut_size.y,
+                    depth_or_array_layers: 1,
+                },
+                mip_level_count: 1,
+                sample_count: 1,
+                dimension: TextureDimension::D2,
+                format: TextureFormat::Rgba16Float,
+                usage: TextureUsages::STORAGE_BINDING | TextureUsages::TEXTURE_BINDING,
+                view_formats: &[],
+            },
+        );
+
+        let multiscattering_lut = texture_cache.get(
+            &render_device,
+            TextureDescriptor {
+                label: Some("multiscattering_lut"),
+                size: Extent3d {
+                    width: lut_settings.multiscattering_lut_size.x,
+                    height: lut_settings.multiscattering_lut_size.y,
+                    depth_or_array_layers: 1,
+                },
+                mip_level_count: 1,
+                sample_count: 1,
+                dimension: TextureDimension::D2,
+                format: TextureFormat::Rgba16Float,
+                usage: TextureUsages::STORAGE_BINDING | TextureUsages::TEXTURE_BINDING,
+                view_formats: &[],
+            },
+        );
+
+        let sky_view_lut = texture_cache.get(
+            &render_device,
+            TextureDescriptor {
+                label: Some("sky_view_lut"),
+                size: Extent3d {
+                    width: lut_settings.sky_view_lut_size.x,
+                    height: lut_settings.sky_view_lut_size.y,
+                    depth_or_array_layers: 1,
+                },
+                mip_level_count: 1,
+                sample_count: 1,
+                dimension: TextureDimension::D2,
+                format: TextureFormat::Rgba16Float,
+                usage: TextureUsages::STORAGE_BINDING | TextureUsages::TEXTURE_BINDING,
+                view_formats: &[],
+            },
+        );
+
+        let aerial_view_lut = texture_cache.get(
+            &render_device,
+            TextureDescriptor {
+                label: Some("aerial_view_lut"),
+                size: Extent3d {
+                    width: lut_settings.aerial_view_lut_size.x,
+                    height: lut_settings.aerial_view_lut_size.y,
+                    depth_or_array_layers: lut_settings.aerial_view_lut_size.z,
+                },
+                mip_level_count: 1,
+                sample_count: 1,
+                dimension: TextureDimension::D3,
+                format: TextureFormat::Rgba16Float,
+                usage: TextureUsages::STORAGE_BINDING | TextureUsages::TEXTURE_BINDING,
+                view_formats: &[],
+            },
+        );
+
+        commands.entity(entity).insert({
+            AtmosphereTextures {
+                transmittance_lut,
+                multiscattering_lut,
+                sky_view_lut,
+                aerial_view_lut,
+            }
+        });
+    }
+}
+
+pub fn prepare_atmosphere_transforms(
+    views: Query<(Entity, &ExtractedView), (With<Atmosphere>, With<Camera3d>)>,
+    render_device: Res<RenderDevice>,
+    render_queue: Res<RenderQueue>,
+    mut atmo_uniforms: ResMut<AtmosphereTransforms>,
+    mut commands: Commands,
+) {
+    let atmo_count = views.iter().len();
+    let Some(mut writer) =
+        atmo_uniforms
+            .uniforms
+            .get_writer(atmo_count, &render_device, &render_queue)
+    else {
+        return;
+    };
+
+    for (entity, view) in &views {
+        let world_from_view = view.world_from_view.compute_matrix();
+        let camera_z = world_from_view.z_axis.truncate();
+        let camera_y = world_from_view.y_axis.truncate();
+        let atmo_z = camera_z
+            .with_y(0.0)
+            .try_normalize()
+            .unwrap_or_else(|| camera_y.with_y(0.0).normalize());
+        let atmo_y = Vec3::Y;
+        let atmo_x = atmo_y.cross(atmo_z).normalize();
+        let world_from_atmosphere = Mat4::from_cols(
+            atmo_x.extend(0.0),
+            atmo_y.extend(0.0),
+            atmo_z.extend(0.0),
+            world_from_view.w_axis,
+        );
+
+        let atmosphere_from_world = world_from_atmosphere.inverse();
+
+        commands.entity(entity).insert(AtmosphereTransformsOffset {
+            index: writer.write(&AtmosphereTransform {
+                world_from_atmosphere,
+                atmosphere_from_world,
+            }),
+        });
+    }
+}
+
+pub fn prepare_atmosphere_bind_groups(
+    views: Query<
+        (Entity, &AtmosphereTextures, &ViewDepthTexture, &Msaa),
+        (With<Camera3d>, With<Atmosphere>),
+    >,
+    render_device: Res<RenderDevice>,
+    layouts: Res<AtmosphereBindGroupLayouts>,
+    render_sky_layouts: Res<RenderSkyBindGroupLayouts>,
+    samplers: Res<AtmosphereSamplers>,
+    view_uniforms: Res<ViewUniforms>,
+    lights_uniforms: Res<LightMeta>,
+    atmosphere_transforms: Res<AtmosphereTransforms>,
+    atmosphere_uniforms: Res<ComponentUniforms<Atmosphere>>,
+    settings_uniforms: Res<ComponentUniforms<AtmosphereSettings>>,
+
+    mut commands: Commands,
+) {
+    if views.iter().len() == 0 {
+        return;
+    }
+
+    let atmosphere_binding = atmosphere_uniforms
+        .binding()
+        .expect("Failed to prepare atmosphere bind groups. Atmosphere uniform buffer missing");
+
+    let transforms_binding = atmosphere_transforms
+        .uniforms()
+        .binding()
+        .expect("Failed to prepare atmosphere bind groups. Atmosphere transforms buffer missing");
+
+    let settings_binding = settings_uniforms.binding().expect(
+        "Failed to prepare atmosphere bind groups. AtmosphereSettings uniform buffer missing",
+    );
+
+    let view_binding = view_uniforms
+        .uniforms
+        .binding()
+        .expect("Failed to prepare atmosphere bind groups. View uniform buffer missing");
+
+    let lights_binding = lights_uniforms
+        .view_gpu_lights
+        .binding()
+        .expect("Failed to prepare atmosphere bind groups. Lights uniform buffer missing");
+
+    for (entity, textures, view_depth_texture, msaa) in &views {
+        let transmittance_lut = render_device.create_bind_group(
+            "transmittance_lut_bind_group",
+            &layouts.transmittance_lut,
+            &BindGroupEntries::with_indices((
+                (0, atmosphere_binding.clone()),
+                (1, settings_binding.clone()),
+                (13, &textures.transmittance_lut.default_view),
+            )),
+        );
+
+        let multiscattering_lut = render_device.create_bind_group(
+            "multiscattering_lut_bind_group",
+            &layouts.multiscattering_lut,
+            &BindGroupEntries::with_indices((
+                (0, atmosphere_binding.clone()),
+                (1, settings_binding.clone()),
+                (5, &textures.transmittance_lut.default_view),
+                (6, &samplers.transmittance_lut),
+                (13, &textures.multiscattering_lut.default_view),
+            )),
+        );
+
+        let sky_view_lut = render_device.create_bind_group(
+            "sky_view_lut_bind_group",
+            &layouts.sky_view_lut,
+            &BindGroupEntries::with_indices((
+                (0, atmosphere_binding.clone()),
+                (1, settings_binding.clone()),
+                (2, transforms_binding.clone()),
+                (3, view_binding.clone()),
+                (4, lights_binding.clone()),
+                (5, &textures.transmittance_lut.default_view),
+                (6, &samplers.transmittance_lut),
+                (7, &textures.multiscattering_lut.default_view),
+                (8, &samplers.multiscattering_lut),
+                (13, &textures.sky_view_lut.default_view),
+            )),
+        );
+
+        let aerial_view_lut = render_device.create_bind_group(
+            "sky_view_lut_bind_group",
+            &layouts.aerial_view_lut,
+            &BindGroupEntries::with_indices((
+                (0, atmosphere_binding.clone()),
+                (1, settings_binding.clone()),
+                (3, view_binding.clone()),
+                (4, lights_binding.clone()),
+                (5, &textures.transmittance_lut.default_view),
+                (6, &samplers.transmittance_lut),
+                (7, &textures.multiscattering_lut.default_view),
+                (8, &samplers.multiscattering_lut),
+                (13, &textures.aerial_view_lut.default_view),
+            )),
+        );
+
+        let render_sky = render_device.create_bind_group(
+            "render_sky_bind_group",
+            if *msaa == Msaa::Off {
+                &render_sky_layouts.render_sky
+            } else {
+                &render_sky_layouts.render_sky_msaa
+            },
+            &BindGroupEntries::with_indices((
+                (0, atmosphere_binding.clone()),
+                (1, settings_binding.clone()),
+                (2, transforms_binding.clone()),
+                (3, view_binding.clone()),
+                (4, lights_binding.clone()),
+                (5, &textures.transmittance_lut.default_view),
+                (6, &samplers.transmittance_lut),
+                (9, &textures.sky_view_lut.default_view),
+                (10, &samplers.sky_view_lut),
+                (11, &textures.aerial_view_lut.default_view),
+                (12, &samplers.aerial_view_lut),
+                (13, view_depth_texture.view()),
+            )),
+        );
+
+        commands.entity(entity).insert(AtmosphereBindGroups {
+            transmittance_lut,
+            multiscattering_lut,
+            sky_view_lut,
+            aerial_view_lut,
+            render_sky,
+        });
+    }
+}
