+++ crates/bevy_render_3d/src/atmosphere/resources.rs
@@ -1,27 +1,23 @@
-use bevy_core_pipeline::{
-    core_3d::Camera3d, fullscreen_vertex_shader::fullscreen_shader_vertex_state,
-};
+use bevy_core_pipeline::fullscreen_vertex_shader::fullscreen_shader_vertex_state;
 use bevy_ecs::{
     component::Component,
-    entity::Entity,
-    query::With,
     resource::Resource,
-    system::{Commands, Query, Res, ResMut},
     world::{FromWorld, World},
 };
-use bevy_math::{Mat4, Vec3};
+use bevy_math::Mat4;
 use bevy_render::{
-    camera::Camera,
-    extract_component::ComponentUniforms,
     render_resource::{binding_types::*, *},
-    renderer::{RenderDevice, RenderQueue},
-    texture::{CachedTexture, TextureCache},
-    view::{ExtractedView, Msaa, ViewDepthTexture, ViewUniform, ViewUniforms},
+    renderer::RenderDevice,
+    texture::CachedTexture,
+    view::ViewUniform,
 };
 
-use crate::{GpuLights, LightMeta};
+use crate::GpuLights;
 
-use super::{shaders, Atmosphere, AtmosphereSettings};
+use super::{
+    plugin::{AERIAL_VIEW_LUT, MULTISCATTERING_LUT, RENDER_SKY, SKY_VIEW_LUT, TRANSMITTANCE_LUT},
+    Atmosphere, AtmosphereSettings,
+};
 
 #[derive(Resource)]
 pub(crate) struct AtmosphereBindGroupLayouts {
@@ -273,7 +269,7 @@ impl FromWorld for AtmosphereLutPipelines {
             label: Some("transmittance_lut_pipeline".into()),
             layout: vec![layouts.transmittance_lut.clone()],
             push_constant_ranges: vec![],
-            shader: shaders::TRANSMITTANCE_LUT,
+            shader: TRANSMITTANCE_LUT,
             shader_defs: vec![],
             entry_point: "main".into(),
             zero_initialize_workgroup_memory: false,
@@ -284,7 +280,7 @@ impl FromWorld for AtmosphereLutPipelines {
                 label: Some("multi_scattering_lut_pipeline".into()),
                 layout: vec![layouts.multiscattering_lut.clone()],
                 push_constant_ranges: vec![],
-                shader: shaders::MULTISCATTERING_LUT,
+                shader: MULTISCATTERING_LUT,
                 shader_defs: vec![],
                 entry_point: "main".into(),
                 zero_initialize_workgroup_memory: false,
@@ -294,7 +290,7 @@ impl FromWorld for AtmosphereLutPipelines {
             label: Some("sky_view_lut_pipeline".into()),
             layout: vec![layouts.sky_view_lut.clone()],
             push_constant_ranges: vec![],
-            shader: shaders::SKY_VIEW_LUT,
+            shader: SKY_VIEW_LUT,
             shader_defs: vec![],
             entry_point: "main".into(),
             zero_initialize_workgroup_memory: false,
@@ -304,7 +300,7 @@ impl FromWorld for AtmosphereLutPipelines {
             label: Some("aerial_view_lut_pipeline".into()),
             layout: vec![layouts.aerial_view_lut.clone()],
             push_constant_ranges: vec![],
-            shader: shaders::AERIAL_VIEW_LUT,
+            shader: AERIAL_VIEW_LUT,
             shader_defs: vec![],
             entry_point: "main".into(),
             zero_initialize_workgroup_memory: false,
@@ -359,7 +355,7 @@ impl SpecializedRenderPipeline for RenderSkyBindGroupLayouts {
             },
             zero_initialize_workgroup_memory: false,
             fragment: Some(FragmentState {
-                shader: shaders::RENDER_SKY.clone(),
+                shader: RENDER_SKY.clone(),
                 shader_defs,
                 entry_point: "main".into(),
                 targets: vec![Some(ColorTargetState {
@@ -383,26 +379,6 @@ impl SpecializedRenderPipeline for RenderSkyBindGroupLayouts {
     }
 }
 
-pub(super) fn queue_render_sky_pipelines(
-    views: Query<(Entity, &Camera, &Msaa), With<Atmosphere>>,
-    pipeline_cache: Res<PipelineCache>,
-    layouts: Res<RenderSkyBindGroupLayouts>,
-    mut specializer: ResMut<SpecializedRenderPipelines<RenderSkyBindGroupLayouts>>,
-    mut commands: Commands,
-) {
-    for (entity, camera, msaa) in &views {
-        let id = specializer.specialize(
-            &pipeline_cache,
-            &layouts,
-            RenderSkyPipelineKey {
-                msaa_samples: msaa.samples(),
-                hdr: camera.hdr,
-            },
-        );
-        commands.entity(entity).insert(RenderSkyPipelineId(id));
-    }
-}
-
 #[derive(Component)]
 pub struct AtmosphereTextures {
     pub transmittance_lut: CachedTexture,
@@ -411,99 +387,9 @@ pub struct AtmosphereTextures {
     pub aerial_view_lut: CachedTexture,
 }
 
-pub(super) fn prepare_atmosphere_textures(
-    views: Query<(Entity, &AtmosphereSettings), With<Atmosphere>>,
-    render_device: Res<RenderDevice>,
-    mut texture_cache: ResMut<TextureCache>,
-    mut commands: Commands,
-) {
-    for (entity, lut_settings) in &views {
-        let transmittance_lut = texture_cache.get(
-            &render_device,
-            TextureDescriptor {
-                label: Some("transmittance_lut"),
-                size: Extent3d {
-                    width: lut_settings.transmittance_lut_size.x,
-                    height: lut_settings.transmittance_lut_size.y,
-                    depth_or_array_layers: 1,
-                },
-                mip_level_count: 1,
-                sample_count: 1,
-                dimension: TextureDimension::D2,
-                format: TextureFormat::Rgba16Float,
-                usage: TextureUsages::STORAGE_BINDING | TextureUsages::TEXTURE_BINDING,
-                view_formats: &[],
-            },
-        );
-
-        let multiscattering_lut = texture_cache.get(
-            &render_device,
-            TextureDescriptor {
-                label: Some("multiscattering_lut"),
-                size: Extent3d {
-                    width: lut_settings.multiscattering_lut_size.x,
-                    height: lut_settings.multiscattering_lut_size.y,
-                    depth_or_array_layers: 1,
-                },
-                mip_level_count: 1,
-                sample_count: 1,
-                dimension: TextureDimension::D2,
-                format: TextureFormat::Rgba16Float,
-                usage: TextureUsages::STORAGE_BINDING | TextureUsages::TEXTURE_BINDING,
-                view_formats: &[],
-            },
-        );
-
-        let sky_view_lut = texture_cache.get(
-            &render_device,
-            TextureDescriptor {
-                label: Some("sky_view_lut"),
-                size: Extent3d {
-                    width: lut_settings.sky_view_lut_size.x,
-                    height: lut_settings.sky_view_lut_size.y,
-                    depth_or_array_layers: 1,
-                },
-                mip_level_count: 1,
-                sample_count: 1,
-                dimension: TextureDimension::D2,
-                format: TextureFormat::Rgba16Float,
-                usage: TextureUsages::STORAGE_BINDING | TextureUsages::TEXTURE_BINDING,
-                view_formats: &[],
-            },
-        );
-
-        let aerial_view_lut = texture_cache.get(
-            &render_device,
-            TextureDescriptor {
-                label: Some("aerial_view_lut"),
-                size: Extent3d {
-                    width: lut_settings.aerial_view_lut_size.x,
-                    height: lut_settings.aerial_view_lut_size.y,
-                    depth_or_array_layers: lut_settings.aerial_view_lut_size.z,
-                },
-                mip_level_count: 1,
-                sample_count: 1,
-                dimension: TextureDimension::D3,
-                format: TextureFormat::Rgba16Float,
-                usage: TextureUsages::STORAGE_BINDING | TextureUsages::TEXTURE_BINDING,
-                view_formats: &[],
-            },
-        );
-
-        commands.entity(entity).insert({
-            AtmosphereTextures {
-                transmittance_lut,
-                multiscattering_lut,
-                sky_view_lut,
-                aerial_view_lut,
-            }
-        });
-    }
-}
-
 #[derive(Resource, Default)]
 pub struct AtmosphereTransforms {
-    uniforms: DynamicUniformBuffer<AtmosphereTransform>,
+    pub uniforms: DynamicUniformBuffer<AtmosphereTransform>,
 }
 
 impl AtmosphereTransforms {
@@ -515,63 +401,19 @@ impl AtmosphereTransforms {
 
 #[derive(ShaderType)]
 pub struct AtmosphereTransform {
-    world_from_atmosphere: Mat4,
-    atmosphere_from_world: Mat4,
+    pub world_from_atmosphere: Mat4,
+    pub atmosphere_from_world: Mat4,
 }
 
 #[derive(Component)]
 pub struct AtmosphereTransformsOffset {
-    index: u32,
+    pub(crate) index: u32,
 }
 
 impl AtmosphereTransformsOffset {
     #[inline]
     pub fn index(&self) -> u32 {
         self.index
-    }
-}
-
-pub(super) fn prepare_atmosphere_transforms(
-    views: Query<(Entity, &ExtractedView), (With<Atmosphere>, With<Camera3d>)>,
-    render_device: Res<RenderDevice>,
-    render_queue: Res<RenderQueue>,
-    mut atmo_uniforms: ResMut<AtmosphereTransforms>,
-    mut commands: Commands,
-) {
-    let atmo_count = views.iter().len();
-    let Some(mut writer) =
-        atmo_uniforms
-            .uniforms
-            .get_writer(atmo_count, &render_device, &render_queue)
-    else {
-        return;
-    };
-
-    for (entity, view) in &views {
-        let world_from_view = view.world_from_view.compute_matrix();
-        let camera_z = world_from_view.z_axis.truncate();
-        let camera_y = world_from_view.y_axis.truncate();
-        let atmo_z = camera_z
-            .with_y(0.0)
-            .try_normalize()
-            .unwrap_or_else(|| camera_y.with_y(0.0).normalize());
-        let atmo_y = Vec3::Y;
-        let atmo_x = atmo_y.cross(atmo_z).normalize();
-        let world_from_atmosphere = Mat4::from_cols(
-            atmo_x.extend(0.0),
-            atmo_y.extend(0.0),
-            atmo_z.extend(0.0),
-            world_from_view.w_axis,
-        );
-
-        let atmosphere_from_world = world_from_atmosphere.inverse();
-
-        commands.entity(entity).insert(AtmosphereTransformsOffset {
-            index: writer.write(&AtmosphereTransform {
-                world_from_atmosphere,
-                atmosphere_from_world,
-            }),
-        });
     }
 }
 
@@ -582,137 +424,4 @@ pub(crate) struct AtmosphereBindGroups {
     pub sky_view_lut: BindGroup,
     pub aerial_view_lut: BindGroup,
     pub render_sky: BindGroup,
-}
-
-pub(super) fn prepare_atmosphere_bind_groups(
-    views: Query<
-        (Entity, &AtmosphereTextures, &ViewDepthTexture, &Msaa),
-        (With<Camera3d>, With<Atmosphere>),
-    >,
-    render_device: Res<RenderDevice>,
-    layouts: Res<AtmosphereBindGroupLayouts>,
-    render_sky_layouts: Res<RenderSkyBindGroupLayouts>,
-    samplers: Res<AtmosphereSamplers>,
-    view_uniforms: Res<ViewUniforms>,
-    lights_uniforms: Res<LightMeta>,
-    atmosphere_transforms: Res<AtmosphereTransforms>,
-    atmosphere_uniforms: Res<ComponentUniforms<Atmosphere>>,
-    settings_uniforms: Res<ComponentUniforms<AtmosphereSettings>>,
-
-    mut commands: Commands,
-) {
-    if views.iter().len() == 0 {
-        return;
-    }
-
-    let atmosphere_binding = atmosphere_uniforms
-        .binding()
-        .expect("Failed to prepare atmosphere bind groups. Atmosphere uniform buffer missing");
-
-    let transforms_binding = atmosphere_transforms
-        .uniforms()
-        .binding()
-        .expect("Failed to prepare atmosphere bind groups. Atmosphere transforms buffer missing");
-
-    let settings_binding = settings_uniforms.binding().expect(
-        "Failed to prepare atmosphere bind groups. AtmosphereSettings uniform buffer missing",
-    );
-
-    let view_binding = view_uniforms
-        .uniforms
-        .binding()
-        .expect("Failed to prepare atmosphere bind groups. View uniform buffer missing");
-
-    let lights_binding = lights_uniforms
-        .view_gpu_lights
-        .binding()
-        .expect("Failed to prepare atmosphere bind groups. Lights uniform buffer missing");
-
-    for (entity, textures, view_depth_texture, msaa) in &views {
-        let transmittance_lut = render_device.create_bind_group(
-            "transmittance_lut_bind_group",
-            &layouts.transmittance_lut,
-            &BindGroupEntries::with_indices((
-                (0, atmosphere_binding.clone()),
-                (1, settings_binding.clone()),
-                (13, &textures.transmittance_lut.default_view),
-            )),
-        );
-
-        let multiscattering_lut = render_device.create_bind_group(
-            "multiscattering_lut_bind_group",
-            &layouts.multiscattering_lut,
-            &BindGroupEntries::with_indices((
-                (0, atmosphere_binding.clone()),
-                (1, settings_binding.clone()),
-                (5, &textures.transmittance_lut.default_view),
-                (6, &samplers.transmittance_lut),
-                (13, &textures.multiscattering_lut.default_view),
-            )),
-        );
-
-        let sky_view_lut = render_device.create_bind_group(
-            "sky_view_lut_bind_group",
-            &layouts.sky_view_lut,
-            &BindGroupEntries::with_indices((
-                (0, atmosphere_binding.clone()),
-                (1, settings_binding.clone()),
-                (2, transforms_binding.clone()),
-                (3, view_binding.clone()),
-                (4, lights_binding.clone()),
-                (5, &textures.transmittance_lut.default_view),
-                (6, &samplers.transmittance_lut),
-                (7, &textures.multiscattering_lut.default_view),
-                (8, &samplers.multiscattering_lut),
-                (13, &textures.sky_view_lut.default_view),
-            )),
-        );
-
-        let aerial_view_lut = render_device.create_bind_group(
-            "sky_view_lut_bind_group",
-            &layouts.aerial_view_lut,
-            &BindGroupEntries::with_indices((
-                (0, atmosphere_binding.clone()),
-                (1, settings_binding.clone()),
-                (3, view_binding.clone()),
-                (4, lights_binding.clone()),
-                (5, &textures.transmittance_lut.default_view),
-                (6, &samplers.transmittance_lut),
-                (7, &textures.multiscattering_lut.default_view),
-                (8, &samplers.multiscattering_lut),
-                (13, &textures.aerial_view_lut.default_view),
-            )),
-        );
-
-        let render_sky = render_device.create_bind_group(
-            "render_sky_bind_group",
-            if *msaa == Msaa::Off {
-                &render_sky_layouts.render_sky
-            } else {
-                &render_sky_layouts.render_sky_msaa
-            },
-            &BindGroupEntries::with_indices((
-                (0, atmosphere_binding.clone()),
-                (1, settings_binding.clone()),
-                (2, transforms_binding.clone()),
-                (3, view_binding.clone()),
-                (4, lights_binding.clone()),
-                (5, &textures.transmittance_lut.default_view),
-                (6, &samplers.transmittance_lut),
-                (9, &textures.sky_view_lut.default_view),
-                (10, &samplers.sky_view_lut),
-                (11, &textures.aerial_view_lut.default_view),
-                (12, &samplers.aerial_view_lut),
-                (13, view_depth_texture.view()),
-            )),
-        );
-
-        commands.entity(entity).insert(AtmosphereBindGroups {
-            transmittance_lut,
-            multiscattering_lut,
-            sky_view_lut,
-            aerial_view_lut,
-            render_sky,
-        });
-    }
 }
